Add IDs to well blocks on settings level.
Also added WellBlock struct on settings level which is now used in
place of a simple integer coordinate to indicate positions where
applicable (i.e. a completion now has a parent WellBlock object
with its location, instead of having the location only). Existing
code should be updated to take advantage of this.