Model initialization complete; Redid control variable handling.
The entire model is now initialized from the Model constructor.
Wrote tests for the Model and Well classes.
Wrote a common test fixture to be used for all tests using settings and variables.
Added getters for the well class.
Modified the driver file to include one control for each time step at which it is variable. This must always be done from now on.