Renamed variables in the model.
# What
Variables in the Model library are now generally called Properties.

IntegerVariable is now called DiscreteProperty and RealVariable is now called ContinousProperty.

# Why
This is to better separate properties and general and properties that are variable.

This will become more important when the Optimization library is implemented, where variables will be variables and properties will not be mentioned (at least not in the same respect as in the Model).

Overall, this facilitates the conceptual separation between the properties of a model and the more primitive variables in the Optimizer.