Improve variable uniqueness constraint.
Controls are now also unique.

`#` is now used for the automatically applied suffixes.

Added check in Model to ensure that all variables are unique.

Getting variables by name from the varaible property conainer now ignores the automatically applied suffixes. This was done in order to make the constraint definitions work.