Placed linebreaks in less "random" locations.
The length of each line doesn'r really matter for markdown code. Single breaks are removed and double brakes are taken as paragraph separators when they are parsed anyway. Thus these linebreaks are really only seen when editing the files, in which case an overuse of linebreaks can be annoying when editiing lines (removing or inserting new line breaks every time you write a word), and they mark an unnecessary number of lines as changed in git. Line breaks are, in my oppinion best placed only after punctuation. This makes the 'gitting' easier later. If you want narrower columns of text, you can activate soft wrapping in the editor.