Get Cell geometry and properties separately.
Cell properties, such as permabilities and porosity, are no longer retrieved when GetCell/GetGridCell is called. Grid::FillCellProperties(&Cell) should now be used to get the non-geometric properties of cells.

This was done partly because the old function crashed when one attempted to retrieve information about inactive cells (which have geometry but no other properties); and partly to speed things up: permeabilities and porosities are only needed when calculating the well index, while the geometry is also needed in the constraints, and to find which cells are intersected by a well spline.