Implemented a `MaxMinutes` setting for simulations.
This new field specifies a maximum number of minutes a simulator is allowed to run. This value will be used when a normal timeout value cannot be determined (e.g. in the first iteration). It is needed to avoid severe hangups in the first iteration for large reservoirs, particularly when used with a stochastic optimizer and a large number of workers.