Added logging of the current runner state
The current state of the runner (in particular Workers) is now written temporarily to a separate file. This file indicates whether the well index has been calculated (and when the calculation started) and whether the simulation has finished (and when it started).

The intention of this is to make it easier to identify problematic cases, as they are normally not logged until simulation has finished.