Matrix vs Fracture active cells
The code makes the distiction between fracture and matrix grids when
checking whether a cell is active. The number of active cells and the
active status methods will have "matrix" or "fracture" in their names to
signal on which grid they work on. The IsCellActive method will check
whether a cell is active on any of the grids.