Matrix vs Fracture cell data
1. Removed bug in the use of IsCellActive method:
The method IsCellActive takes a global index as argument. The previous
version was using it with the active index, that would have given
unexpected results.
2. Added the flags that will record the active state of a cell on both
matrix and fracture grids
3. The poro, permx, permy and permz members are now vectors that store
values for both grids. If a cell is active on both grids the value for
the matrix grid is stored first.