Added new class for handling file paths.
The new class provides automatic checks on file paths, and allows you to check whether or not a path has been set.

It eases the setting of paths in the RuntimeSettings, and makes it easier to extend. It will also be passed into the Settings and Simulation classes to make it easier to keep things in sync.