Fix an issue regarding the first imported well block.
The first imported well block for each well gets wildly wrong IJK indices. Previously, I tried just "guessing" the values from either the well heel or from the second well block, but this was too often wrong. Instead, the first block is now just deleted if issues are detected. This is not ideal, but it causes a smaller error and fewer issues than any other solution I can think of.