Implement passing of static RNG seed.
An RNG (Random Number Generator) seed can now be set in the json driver file (under Optimization.Parameters). This will be passed to both optimization algorithms and the EnsembleHelper.

This allows for reproducible results with stochastic algorithms.

If no seed is passed, the current (UNIX) time will be set as the seed by default.