Packer variables now are now placement relative to well length.
The variable for placement of packers is now the relative placement along the length of the well. All packers are now implicitly constrained to the range [0, 1]. This eases the application of constraints in the optimizer. This means that a placement of 0.5 means that the packer is placed half-way down the wellbore.