Add some "memory balancing" for ensemble runs
In EnsembleHelper: assigning realizations to specific workers. A realization will be evaluated by a worker it has already been assigned to (given that the worker is currently available). This means that, generally, each worker will not have to deal with all realizations. This reduces both the memory requirement and the disk space used.

In Overseer: added method to get list of free Workers and added ability to send a case to a specific worker.