Fix issue with step lengths
Previously, at successful iterations, _all_ step lengths would be set to the step length that produced the successful candidate.
This is a problem when running optimizations with AutoScaleStep lengths, as the step lengths along very small dimensions may be set to very  large numbers if a step along a large dimension is successful.