RNGSeed can no longer be set in driver files.
Removed ability to set the RNG seed to be used in an optimizer.
This was done because the RNG seed would only have the desired effect
under very specific conditions: running an algorithm in serial mode on
one specific system. Some algorithms will for some reason not give
reproducible results even under this condition.

**As the setting will in most cases just cause confusion, it has been
removed**. Setting the RNGSeed field in a driver file will cause an
error.

The parameter field in the settings has been kept, as it is
still useful in some unit tests.