FieldOpt can now take a file describing insets for the sim schedule
A path to a file describing a set of snippets/insets to be inserted verbatim into the schedule section may now be provided using the --sim-inset flag.

The feature is thoroughly described in the schedule_insets.h file.

An example of an inset file is provided in examples/ECLIPSE/schedule_inset.txt