Implement compartmentalization through coupling of ICV size variables.
This feature allows for grouping of individual ICVs into "compartments". All ICVs in a compartment will have their valve size represented by _one_ variable which will be applied to all ICVs in the compartment. This allows for creation of compartments without modelling of packers/annulus. It results in faster, more stable and less error prone simulations, but it introduces a very small inaccuracy.