Support post-simulation script execution
FieldOpt is now able to execute a shell script after each simulation.
This script may do whatever (FieldOpt does not care).

For now, this feature is used as follows:

1. In the driver file, in the simulation section, add `"UsePostSimScript": true`.
This will make FieldOpt search for and (if it is found) execute a script named `FO_POSTSIM.sh`
located in the simulator directory (i.e. the same directory that contains the .DATA file).
2. A file named `FO_POSTSIM.sh` must be placed in the same directory as the .DATA simulator file.
If running with an ensemble, one such file must be placed in each realization directory.

Note that FieldOpt should not crash if no such script is found: a warning will be printed.