Enable using ACTIONX with well controls
ACTIONX-style driver file writing may now be used
with control optimization.
This means that control optimization may now be
performed without specifying anything but a wells
name and the control variables.

Do note that in order to avoid problems with controls
being specified for the beginning of simulations,
controls specified in FieldOpt for day 0 will actually
be applied on day 1.